requirements.txt
uvicorn~=0.22.0
PyYAML~=6.0
Jinja2~=3.1.2
loguru~=0.7.0
fastapi~=0.96.0